# Feat: Multiple perf improvements and `viztracer` performance profiling