--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,6 @@
+# Viztracer log files
+viztracer_report.json
+
 # Packaged docs
 docs/*.zip