--- a/examples/run_perf_test_reads.py
+++ b/examples/run_perf_test_reads.py
@@ -4,6 +4,8 @@ Simple script to get performance profile of read throughput.
 
 This script accepts a single argument `-e=SCALE` as a power of 10.
 
+-e=2 is equivalent to 500 records.
+-e=3 is equivalent to 5_000 records.
 -e=4 is equivalent to 50_000 records.
 -e=5 is equivalent to 500_000 records.
 -e=6 is equivalent to 5_000_000 records.
@@ -11,11 +13,26 @@ This script accepts a single argument `-e=SCALE` as a power of 10.
 Use smaller values of `e` (2-3) to understand read and overhead costs.
 Use larger values of `e` (4-5) to understand write throughput at scale.
 
-For performance profiling:
+For performance profiling, use `viztracer` to generate a flamegraph:
 ```
 poetry run viztracer --open -- ./examples/run_perf_test_reads.py -e=3
 poetry run viztracer --open -- ./examples/run_perf_test_reads.py -e=5
 ```
+
+To run without profiling, prefix script name with `poetry run python`:
+```
+# Run with 5_000 records
+poetry run python ./examples/run_perf_test_reads.py -e=3
+# Run with 500_000 records
+poetry run python ./examples/run_perf_test_reads.py -e=5
+
+# Load 5_000 records to Snowflake
+poetry run python ./examples/run_perf_test_reads.py -e=3 --cache=snowflake
+
+# Load 5_000 records to BigQuery
+poetry run python ./examples/run_perf_test_reads.py -e=3 --cache=bigquery
+```
+
 """
 
 from __future__ import annotations