--- a/pyproject.toml
+++ b/pyproject.toml
@@ -67,6 +67,7 @@ pytest-mock = "^3.14.0"
 poethepoet = "^0.26.1"
 coverage = "^7.5.1"
 pytest-timeout = "^2.3.1"
+viztracer = "^0.16.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]