<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Introduced default configuration settings for VizTracer to enhance program execution tracing and visualization.
  - Added scripts for profiling performance of Snowflake cache and read throughput.

- **Enhancements**
  - Improved schema handling with force refresh option for enhanced data integrity.
  - Updated batch size limit for file writing to improve performance.
  - Enabled parallel file uploads in Snowflake using thread pools.
  - Applied caching to the normalize method for better performance.
  - Enhanced time calculation precision in progress tracking.

- **Bug Fixes**
  - Fixed a bug in dataset name handling for BigQuery integration tests.

- **Chores**
  - Updated `.gitignore` to exclude VizTracer log files and packaged documentation zips.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

## General Perf Improvements (Read-to-DuckDB)

Before (15K records per second):

![image](https://github.com/user-attachments/assets/b335033a-f810-4579-a92f-8e86083d5495)

After (23K records per second):

![image](https://github.com/user-attachments/assets/77676fed-4d3b-44db-b131-3ed51e8244d3)

## Snowflake Perf Improvements (50K records)

Before (load takes 32 seconds):

![image](https://github.com/user-attachments/assets/30d9f5b1-2921-4d4e-8f57-19a8e40dff9a)

After (load takes 25 seconds):

![image](https://github.com/user-attachments/assets/6445673b-a0ff-4ff1-a965-fa994183cf47)

## Snowflake Perf Improvements (500K records)

Before (load takes 2 min, 43 seconds):

![image](https://github.com/user-attachments/assets/ea38fc39-878d-4cf9-8057-cd3f8d55cc8a)

After (load takes 25 seconds):

![image](https://github.com/user-attachments/assets/ed326de7-e2f9-49f7-aaf5-063238139290)